fix: phistory
Some arguments on the OpenAPI spec for player service was wrongly marked as required. d1f0a6a28d4db0e0e09d83d1751823063a17bb19 was pushed to fix it however it now force included executorId which is incorrect behaviour as it is now searching for punishments by the command executor only.

This basically reverts that commit.